% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oauni.R
\docType{package}
\name{oauni-package}
\alias{oauni}
\alias{oauni-package}
\title{oauni: Open Access at Universities in Germany}
\description{
Monitoring the uptake of open access in publications from German
  research institutions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://subugoe.github.io/oauni}
  \item \url{https://github.com/subugoe/oauni}
  \item Report bugs at \url{https://github.com/subugoe/oauni/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (\href{https://orcid.org/0000-0002-4703-5388}{ORCID})

Authors:
\itemize{
  \item Najko Jahn \email{najko.jahn@gmail.com} (\href{https://orcid.org/0000-0001-5105-1463}{ORCID})
  \item Anne Hobert \email{hobert@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0003-2429-2995}{ORCID})
}

Other contributors:
\itemize{
  \item Philipp Mayr \email{philippmayr@gmail.com} (\href{https://orcid.org/0000-0002-6656-1658}{ORCID}) [contributor]
  \item Birgit Schmidt \email{bschmidt@sub.uni-goettingen.de} (\href{https://orcid.org/0000-0001-8036-5859}{ORCID}) [contributor]
  \item Niels Taubert \email{post@niels-taubert.de} (\href{https://orcid.org/0000-0002-2357-2648}{ORCID}) [contributor]
  \item Göttingen State and University Library [copyright holder, funder]
  \item Federal Ministry of Education and Research [sponsor, funder]
}

}
\keyword{internal}
